% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_nodes}
\alias{count_nodes}
\title{Search an xml file for a specific named node and count the number of instances}
\usage{
count_nodes(x, xpath)
}
\arguments{
\item{x}{An xml object}

\item{xpath}{string. The node name to search for within the xpath.}
}
\value{
Returns a numeric count value
}
\description{
Search an xml file for a specific named node and count the number of instances
}
