% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_level_utils.R
\name{extract_line_level_data}
\alias{extract_line_level_data}
\title{Open data from a single line metadata table where it's zip or xml format}
\usage{
extract_line_level_data(file)
}
\arguments{
\item{file}{A single row of table metadata extracted using get_timetable_metadata()}
}
\value{
returns a dataframe of information extracted from the given xml or zip url
}
\description{
Open data from a single line metadata table where it's zip or xml format
}
