% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_level_utils.R
\name{extract_service_lookup}
\alias{extract_service_lookup}
\title{Extract service journey and section reference lookup from the provided xml document}
\usage{
extract_service_lookup(i, xml)
}
\arguments{
\item{i}{A vector of node numbers to loop over}

\item{xml}{an xml object containing the relevant nodes}
}
\value{
returns a dataframe of information extracted from the given nodes in the xml
}
\description{
Extract service journey and section reference lookup from the provided xml document
}
