% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_level_utils.R
\name{extract_vehicle_journeys}
\alias{extract_vehicle_journeys}
\title{Extract vehicle-level journey departure times}
\usage{
extract_vehicle_journeys(xml)
}
\arguments{
\item{xml}{an xml object containing the relevant nodes}
}
\value{
returns a dataframe of information extracted from the given nodes in the xml
}
\description{
Extract vehicle-level journey departure times
}
