% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_level_utils.R
\name{extract_vehicle_timing}
\alias{extract_vehicle_timing}
\title{Extract stop-level timing data from vehicle data in the provided xml document}
\usage{
extract_vehicle_timing(i, xml)
}
\arguments{
\item{i}{A vector of node numbers to loop over}

\item{xml}{an xml object containing the relevant nodes}
}
\value{
returns a dataframe of information extracted from the given nodes in the xml
}
\description{
Extract stop-level timing data from vehicle data in the provided xml document
}
