% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timetable_data.R
\name{extract_zip_or_xml}
\alias{extract_zip_or_xml}
\title{Open data from a single line metadata table where it's zip or xml format}
\usage{
extract_zip_or_xml(file)
}
\arguments{
\item{file}{A single row of table metadata extracted using get_timetable_metadata()}
}
\value{
returns a dataframe of information extracted from the given xml or zip url
}
\description{
Open data from a single line metadata table where it's zip or xml format
}
