% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_node_value}
\alias{find_node_value}
\title{Search an xml file for a specific named mode and return the value(s) stored in it}
\usage{
find_node_value(x, xpath)
}
\arguments{
\item{x}{An xml object}

\item{xpath}{string. The node name to search for within the xpath.}
}
\value{
Returns a single string of values from the specified node
}
\description{
Search an xml file for a specific named mode and return the value(s) stored in it
}
