% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timetable_metadata.R
\name{get_timetable_metadata}
\alias{get_timetable_metadata}
\title{Return timetable metadata from the BODS API}
\usage{
get_timetable_metadata(api_key, limit = 25)
}
\arguments{
\item{api_key}{API key for the BODS dataset passed as a string. Can be obtained from \link(https://data.bus-data.dft.gov.uk/api/)}

\item{limit}{integer. Maximum number of records to return for a query. Defaults to 25

To return a lookup of current valid ATCO codes, use the lookup_atco_codes() function.}
}
\value{
Returns a data frame of timetable metadata including links to data from the BODS API.
}
\description{
Return timetable metadata from the BODS API
}
