% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_level_utils.R
\name{line_level_xml}
\alias{line_level_xml}
\title{Pull a table of relevant values from specified nodes in the xml}
\usage{
line_level_xml(x, count = 1, total_count = 1)
}
\arguments{
\item{x}{An xml object}

\item{count}{numeric. Where the xml file is taken from a zip collection, the number file it is. Defaults to 1.}

\item{total_count}{numeric. Where the xml file is taken from a zip collection, the total number of files in the zip. Defaults to 1.}
}
\value{
Returns a table of values extracted from specified nodes of
an xml document
}
\description{
Pull a table of relevant values from specified nodes in the xml
}
