% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noc_lookup.R
\name{noc_lookup}
\alias{noc_lookup}
\title{Lookup between operator names and national operator code (NOC) lookup}
\usage{
noc_lookup()
}
\value{
Returns a data frame of operator names and their corresponding national
operator code from the BODS API.
}
\description{
Lookup between operator names and national operator code (NOC) lookup
}
