% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_null_date}
\alias{not_null_date}
\title{Join together a date value and an associated API string if the value
is not NULL}
\usage{
not_null_date(date, string)
}
\arguments{
\item{date}{R object to check whether it is NULL or not}

\item{string}{API string to append to R object}
}
\description{
Join together a date value and an associated API string if the value
is not NULL
}
