% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_level_utils.R
\name{open_all_xml}
\alias{open_all_xml}
\title{Open every xml file within a zip object and extract data of interest from it using a given function}
\usage{
open_all_xml(url, fun)
}
\arguments{
\item{url}{A url pointing towards a zip object}

\item{fun}{name of a data extracting function to apply to the zip folder}
}
\value{
returns a dataframe of information extracted from xml documents
}
\description{
Open every xml file within a zip object and extract data of interest from it using a given function
}
