% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_level_utils.R
\name{stop_level_xml}
\alias{stop_level_xml}
\title{Extract a tidy dataset of stop-level journey data from the provided xml document}
\usage{
stop_level_xml(xml, count = 1, total_count = 1)
}
\arguments{
\item{xml}{an xml object containing the relevant nodes}

\item{count}{numeric. Where the xml file is taken from a zip collection, the number file it is. Defaults to 1.}

\item{total_count}{numeric. Where the xml file is taken from a zip collection, the total number of files in the zip. Defaults to 1.}
}
\value{
returns a dataframe of information extracted from the given nodes in the xml
}
\description{
Extract a tidy dataset of stop-level journey data from the provided xml document
}
