% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_xml.R
\name{xml_file_counter}
\alias{xml_file_counter}
\title{Count the number of xml files included within a provided metadata dataframe, whether the provided file links are xml or zip}
\usage{
xml_file_counter(timetable_metadata)
}
\arguments{
\item{timetable_metadata}{A table of metadata extracted using get_timetable_metadata()}
}
\value{
returns a numeric vector of number of xml files by row of the provided metadata
}
\description{
Count the number of xml files included within a provided metadata dataframe, whether the provided file links are xml or zip
}
