% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fares_metadata.R
\name{get_fares_metadata}
\alias{get_fares_metadata}
\alias{get_timetable_metadata}
\title{Return fares metadata from the BODS API}
\usage{
get_timetable_metadata(
  api_key = Sys.getenv("BODS_KEY"),
  limit = 25,
  search = NULL
)
}
\arguments{
\item{api_key}{API key for the BODS dataset passed as a string. Can be obtained from \link(https://data.bus-data.dft.gov.uk/api/)}

\item{limit}{integer. Maximum number of records to return for a query. Defaults to 25}

\item{search}{string to search records on; can be a value or partial value to
match the data set name, data set description, organisation name, or admin
area name.}
}
\value{
Returns a data frame of fares metadata including links to data from the BODS API.
}
\description{
Return fares metadata from the BODS API
}
