% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timetable_data.R
\name{get_timetable_data}
\alias{get_timetable_data}
\title{Extract line-level timetable data from all rows of the provided metadata table}
\usage{
get_timetable_data(timetable_metadata)
}
\arguments{
\item{timetable_metadata}{A single row of table metadata extracted using get_timetable_metadata()}
}
\value{
returns list of timetable dataframes, with each dataframe corresponding to a
row on the provided timetable metadata
}
\description{
Extract line-level timetable data from all rows of the provided metadata table
}
\examples{

\dontrun{
#Before running these examples, ensure you have an API key saved
#Return the first 5 results of timetable metadata with no filters
metadata <- get_timetable_metadata(limit = 5)


}

}
