% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timetable_data.R
\name{get_timetable_data}
\alias{get_timetable_data}
\title{Extract timetable data from all rows of the provided metadata table}
\usage{
get_timetable_data(timetable_metadata, level = "line")
}
\arguments{
\item{timetable_metadata}{A single row of table metadata extracted using get_timetable_metadata()}

\item{level}{A string specifying whether data returned should be at the
bus line or individual stop level. Options can be "line" or "stop"}
}
\value{
returns list of timetable dataframes, with each dataframe corresponding to a
row on the provided timetable metadata
}
\description{
Extract timetable data from all rows of the provided metadata table
}
