% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{line_level_xml}
\alias{line_level_xml}
\title{Pull a table of relevant values from specified nodes in the xml}
\usage{
line_level_xml(x)
}
\arguments{
\item{x}{An xml object}
}
\value{
Returns a table of values extracted from specified nodes of
an xml document
}
\description{
Pull a table of relevant values from specified nodes in the xml
}
