% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_null}
\alias{not_null}
\title{Join together a value and an associated API string if the value
is not NULL}
\usage{
not_null(obj, string)
}
\arguments{
\item{obj}{R object to check whether it is NULL or not}

\item{string}{API string to append to R object}
}
\description{
Join together a value and an associated API string if the value
is not NULL
}
