% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{open_all_xml}
\alias{open_all_xml}
\title{Open every xml file within a zip object and extract data of interest from it}
\usage{
open_all_xml(url)
}
\arguments{
\item{x}{A url pointing towards a zip object}
}
\value{
returns a dataframe of information extracted from xml documents
}
\description{
Open every xml file within a zip object and extract data of interest from it
}
